services/admin_service/requirements.txt
# Admin Service Requirements
-r ../../requirements/base.txt

# Admin-specific dependencies
django-extensions>=3.2.0
django-debug-toolbar>=4.0.0